Refactor ConvertedFunction to inherit from ListManipulationFunction
ConvertedFunction has been refactored to avoid code duplication and simplify future changes. Now, instead of inheriting from NativeFunction, it inherits from the new abstract class ListManipulationFunction. All code responsible for list manipulation has been moved to ListManipulationFunction, which should be the parent class for all classes that perform operations on lists.
